Created damage type resistances

+++ Assets/Scripts/Components/BlockComponent.cs
@@ -5,6 +5,6 @@ namespace DapperDino.DamageSystems.Components
     [GenerateAuthoringComponent]
     public struct Block : IComponentData
     {
-        public int Value;
+        public float Value;
     }
 }
